requirements.txt
matplotlib==3.7.1
networkx==3.1
numpy-stl==3.0.1
plotly==5.14.1
torch==2.0.1
torch-geometric==2.3.1
torchvision==0.15.2
trimesh==3.21.7